requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
gunicorn==20.1.0
Pygments==2.12.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
